// front-end-react/src/types/Anotacion.ts
export type Anotacion = {
    idEntidad: number;
    idProceso: number;
    idExpediente: number;
    idFase: string;
    fechaTraslado: Date;
    fechaAsignacion:Date;
    idPrivacidad: number;
    privacidad:string;
    idAnotacion:number;
    anotacion:string;
    idUsuarioRegistro: number;
    usuarioRegistro:string;
    fechaRegistro:Date;
  };